Function did::pallet::dispatchables::delete

pub fn delete<T: Config>(endpoints_to_remove: u32)

Delete a DID from the chain and all information associated with it, after verifying that the delete operation has been signed by the DID subject using the authentication key currently stored on chain.

The referenced DID identifier must be present on chain before the delete operation is evaluated.

After it is deleted, a DID with the same identifier cannot be re-created ever again.

As the result of the deletion, all traces of the DID are removed from the storage, which results in the invalidation of all attestations issued by the DID subject.

The dispatch origin must be a DID origin proxied via the submit_did_call extrinsic.

Emits DidDeleted.

Weight: O(1)

Reads: [Origin Account], Did

Kills: Did entry associated to the DID identifier

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::delete.